semana06/dia6/series/requirements.txt
django==3.2
djangorestframework
django-cors-headers
djongo